requirements.txt
# API access
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# SEC API & Web scraping
sec-api>=0.3.2
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
google-api-python-client>=2.100.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.66.0

seaborn

streamlit
pandas
matplotlib
plotly

fastapi
uvicorn
sqlalchemy
psycopg2-binary